refactor: follow use casting
